requirements.txt
pygame>=2.5.2
greenlet>=3.0.3
gevent>=24.2.1
pillow>=10.3.0
fastapi>=0.110.1
uvicorn>=0.29.0
starlette>=0.37.2
pydantic>=2.7.0
python-jose>=3.3.0
python-multipart>=0.0.9
requests>=2.31.0
numpy>=1.26.4